Case Ajusted and inserted parameter for RawData source

+++ src/MyAnimeList/MyAnimeList/Features/Import/ImportDataHandler.cs
@@ -3,12 +3,10 @@ using CsvHelper.Configuration;
 using MediatR;
 using MyAnimeList.Domain;
 using MyAnimeList.Domain.CsvDomain;
-using Nudes.Paginator.Core;
 using Nudes.Retornator.Core;
 using System.Text;
-using System.Linq;
-using Microsoft.EntityFrameworkCore;
 using Mapster;
+
 
 namespace MyAnimeList.Features.Import;
 
@@ -24,12 +22,25 @@ public class ImportDataHandler : IRequestHandler<ImportDataRequest, ResultOf<boo
     {
         var config = new CsvConfiguration(new System.Globalization.CultureInfo("en-US")) { Delimiter = "," };
 
+        if (request.PathToRawFolder == "string") request.PathToRawFolder = "..\\..\\..\\RawData";
+        var path = request.PathToRawFolder;
+
+
+        if (!File.Exists(Path.Combine(request.PathToRawFolder,"animelist.csv"))
+            || !File.Exists(Path.Combine(request.PathToRawFolder,"anime.csv"))
+            || !File.Exists(Path.Combine(request.PathToRawFolder, "anime_with_synopsis.csv"))
+            || !File.Exists(Path.Combine(request.PathToRawFolder, "rating_complete.csv"))
+            || !File.Exists(Path.Combine(request.PathToRawFolder, "watching_status.csv")))
+        {
+            return new Nudes.Retornator.AspnetCore.Errors.NotFoundError();
+        }
+
         #region AnimeScores
         _context.RemoveRange(_context.AnimeScores.Select(a => a));
-        using (var reader = new StreamReader("..\\..\\..\\RawData\\animelist.csv", Encoding.UTF8))
+        using (var reader = new StreamReader(Path.Combine(path, "animelist.csv"), Encoding.UTF8))
         using (var csv = new CsvReader(reader, config))
         {
-            IEnumerable<animelist> records = csv.GetRecords<animelist>();
+            IEnumerable<AnimeList> records = csv.GetRecords<AnimeList>();
 
             _context.AddRange(records.Select(anime => new AnimeScore
             {
@@ -44,11 +55,11 @@ public class ImportDataHandler : IRequestHandler<ImportDataRequest, ResultOf<boo
 
         #region Animes
         _context.RemoveRange(_context.Animes.Select(a => a));
-        using (var reader = new StreamReader("..\\..\\..\\RawData\\anime.csv", Encoding.UTF8))
+        using (var reader = new StreamReader(Path.Combine(path,"anime.csv"), Encoding.UTF8))
         using (var csv = new CsvReader(reader, config))
         {
             csv.Context.RegisterClassMap<AnimeMap>();
-            IEnumerable<animeraw> records = csv.GetRecords<animeraw>();
+            IEnumerable<AnimeRaw> records = csv.GetRecords<AnimeRaw>();
             _context.AddRange(records.Adapt<IEnumerable<Anime>>());
 
         }
@@ -56,7 +67,7 @@ public class ImportDataHandler : IRequestHandler<ImportDataRequest, ResultOf<boo
 
         #region AnimeWithSynopsis
         _context.RemoveRange(_context.AnimesWithSynopsis.Select(a => a));
-        using (var reader = new StreamReader("..\\..\\..\\RawData\\anime_with_synopsis.csv", Encoding.UTF8))
+        using (var reader = new StreamReader(Path.Combine(path,"anime_with_synopsis.csv"), Encoding.UTF8))
         using (var csv = new CsvReader(reader, config))
         {
             csv.Context.RegisterClassMap<AnimeWithSynopsisMap>();
@@ -67,7 +78,7 @@ public class ImportDataHandler : IRequestHandler<ImportDataRequest, ResultOf<boo
 
         #region RatingFromComplete
         _context.RemoveRange(_context.RatingCompletes.Select(a => a));
-        using (var reader = new StreamReader("..\\..\\..\\RawData\\rating_complete.csv", Encoding.UTF8))
+        using (var reader = new StreamReader(Path.Combine(path,"rating_complete.csv"), Encoding.UTF8))
         using (var csv = new CsvReader(reader, config))
         {
             csv.Context.RegisterClassMap<RatingFromCompleteMap>();
@@ -78,7 +89,7 @@ public class ImportDataHandler : IRequestHandler<ImportDataRequest, ResultOf<boo
 
         #region WatchStatus
         _context.RemoveRange(_context.WatchStatus.Select(a => a));
-        using (var reader = new StreamReader("..\\..\\..\\RawData\\watching_status.csv"))
+        using (var reader = new StreamReader(Path.Combine(path, "watching_status.csv")))
         using (var csv = new CsvReader(reader, config))
         {
             csv.Context.RegisterClassMap<WatchStatusMap>();
